pyproject.toml
[project]
name = "multiagent-delfos-app"
version = "0.1.0"
description = "Multi-agent system for SQL queries and visualization using Microsoft Agent Framework"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Microsoft Agent Framework
    "agent-framework>=0.1.0",
    "agent-framework-azure-ai>=0.1.0",

    # Azure Identity for authentication (az login)
    "azure-identity>=1.19.0",

    # FastAPI & Server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",

    # Configuration
    "python-dotenv>=1.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
